LightingFacadeTest.java
public class LightingFacadeTest {
    private LightingFacade lightingFacade;

    @Before
    public void setUp() {
        lightingFacade = new LightingFacade();
    }

    @Test
    public void testTurnOnAll() {
        lightingFacade.turnOnAll();
        for (String zone : new String[]{"living_room", "kitchen", "bedroom"}) {
            assertTrue("Light in " + zone + " should be on", lightingFacade.getLight(zone).isOn());
        }
    }

    @Test
    public void testTurnOffAll() {
        lightingFacade.turnOnAll();
        lightingFacade.turnOffAll();
        for (String zone : new String[]{"living_room", "kitchen", "bedroom"}) {
            assertFalse("Light in " + zone + " should be off", lightingFacade.getLight(zone).isOn());
        }
    }

    @Test
    public void testSetBrightness() {
        lightingFacade.setBrightness("living_room", 70);
        assertEquals("Brightness in living room should be 70", 70, lightingFacade.getLight("living_room").getBrightness());
    }

    @Test
    public void testSetColor() {
        lightingFacade.setColor("kitchen", "red");
        assertEquals("Color in kitchen should be red", "red", lightingFacade.getLight("kitchen").getColor());
    }

    @Test
    public void testSceneEvening() {
        lightingFacade.sceneEvening();
        assertEquals("Living room color should be warm white", "warm white", lightingFacade.getLight("living_room").getColor());
        assertEquals("Living room brightness should be 50", 50, lightingFacade.getLight("living_room").getBrightness());
        assertEquals("Kitchen color should be soft yellow", "soft yellow", lightingFacade.getLight("kitchen").getColor());
        assertEquals("Kitchen brightness should be 30", 30, lightingFacade.getLight("kitchen").getBrightness());
        assertEquals("Bedroom color should be dim blue", "dim blue", lightingFacade.getLight("bedroom").getColor());
        assertEquals("Bedroom brightness should be 20", 20, lightingFacade.getLight("bedroom").getBrightness());
    }

    @Test
    public void testSceneMovie() {
        lightingFacade.sceneMovie();
        for (String zone : new String[]{"kitchen", "bedroom"}) {
            assertFalse("Light in " + zone + " should be off for movie scene", lightingFacade.getLight(zone).isOn());
        }
        assertEquals("Living room brightness should be 20 for movie scene", 20, lightingFacade.getLight("living_room").getBrightness());
    }

    private void assertEquals(String brightness_in_living_room_should_be_70, int i, int brightness) {
        throw new UnsupportedOperationException("Not supported yet.");
    }

    private void assertEquals(String color_in_kitchen_should_be_red, String red, String color) {
        throw new UnsupportedOperationException("Not supported yet.");
    }

    private void assertFalse(String string, boolean on) {
        throw new UnsupportedOperationException("Not supported yet.");
    }

    private void assertTrue(String string, boolean on) {
        throw new UnsupportedOperationException("Not supported yet.");
    }
}

